Updated removelinks.sh     - Added check for files and directory checking

+++ removelinks.sh
@@ -1,10 +1,23 @@
-!#/usr/bin/bash
+#!/usr/bin/bash
 
-rm -r ~/.vimrc
-rm -r ~/.vim/autoload
-rm -r ~/.config/nvim
-rm -r ~/.zshrc
-rm -r ~/.oh-my-zsh/custom
-rm ~/.aliases.zsh
-rm ~/.lazy-load.zsh
-rm ~/.tmux.conf
+read -p "Have you moved dofiles to a different location?(y/n)" choice
+case $choice in
+    [Yy]*)
+        [[ -f ~/.vimrc ]] && rm ~/.vimrc
+        [[ -d ~/.vim/autoload ]] && rm -r ~/.vim/autoload
+        [[ -d ~/.vim/colors ]] && rm -r ~/.vim/colors
+        [[ -f ~/.vimrc.minimal ]] && rm ~/.vimrc.minimal
+        [[ -d ~/.config/nvim ]] && rm -r ~/.config/nvim
+        [[ -f ~/.zshrc ]] && rm ~/.zshrc
+        [[ -d ~/.oh-my-zsh/custom ]] && rm -r ~/.oh-my-zsh/custom
+        [[ -f ~/.aliases.zsh ]] && rm ~/.aliases.zsh
+        [[ -f ~/.lazy-load.zsh ]] && rm ~/.lazy-load.zsh
+        [[ -f ~/.tmux.conf ]] && rm ~/.tmux.conf
+        ;;
+    [Nn]*)
+        echo -e "Please move the dotfiles to different location to avoid removing it in the version control"
+        ;;
+    *)
+        echo -e "Invalid input!! Exit"
+        ;;
+esac
